Fix deprecation warnings

+++ mindboggle/guts/graph.py
@@ -134,8 +134,8 @@ def weight_graph(Nodes, Indices, Meshes, kernel=rbf_kernel, add_to_graph=True,
             edge_mat = Meshes
         # Augment matrix to contain edge weight in the third column
         weighted_edges = np.asarray([[Indices[np.int(i)], Indices[np.int(j)],
-            kernel(Nodes[Indices[np.int(i)]],
-                   Nodes[Indices[np.int(j)]], sigma)]
+            kernel(Nodes[np.int(Indices[np.int(i)])],
+                   Nodes[np.int(Indices[np.int(j)])], sigma)]
                    for [i, j] in edge_mat])
 
         # Add weights to graph
